tap % api function modified
