+++ package.json
@@ -23,11 +23,11 @@
     "css-loader": "^6.5.1",
     "style-loader": "^3.3.1",
     "webpack": "^5.65.0",
-    "webpack-cli": "^4.9.1"
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-server": "^4.7.2"
   },
   "dependencies": {
     "axios": "^0.24.0",
-    "html-webpack-plugin": "^5.5.0",
-    "webpack-dev-server": "^4.7.2"
+    "html-webpack-plugin": "^5.5.0"
   }
 }
