+++ src/components/SearchResult.js
@@ -1,3 +1,4 @@
+//
 function SearchResult() {
   const wrapper = document.querySelector("#wrapper");
   const searchResult = document.createElement("section");
@@ -6,13 +7,11 @@ function SearchResult() {
   searchResult.id = "searchResult";
   wrapper.appendChild(searchResult);
 
-  return (searchResult.innerHTML = `
-      <div class="users">
-      <span class="user_alphabet"></span>
-      <div class="user_profile">
-        </button>
-      </div>
-    </div>
-      `);
+  {
+    return (searchResult.innerHTML = `
+      <div class="users"></div>
+   
+`);
+  }
 }
 export default SearchResult;
