+++ src/utils/changeTab.js
@@ -1,11 +1,21 @@
 function changeTab(e) {
-  let currentTap = "api";
   const selectedTab = e.target.classList.value;
-  console.log(selectedTab);
+  console.log("main", e.target.classList.value);
   if (selectedTab.includes("api")) {
-    return (currentTap = "api");
+    subbar.innerHTML = `
+            <nav>
+              <button class="subbar_menu api selected">API</button>
+              <button class="subbar_menu local">로컬</button>
+            </nav>
+    `;
   } else if (selectedTab.includes("local")) {
-    return (currentTap = "local");
+    subbar.innerHTML = `
+            <nav>
+            <button class="subbar_menu api">API</button>
+             <button class="subbar_menu local selected">로컬</button>
+            </nav>
+    `;
   }
 }
+
 export default changeTab;
